Rohan | Changed the duration to be of integer type
Conflicts:
	bahmnicore-omod/src/test/java/org/openmrs/module/bahmnicore/web/v1_0/mapper/DrugOrderMapperTest.java

+++ bahmnicore-api/src/test/java/org/bahmni/module/bahmnicore/mapper/builder/DrugOrderBuilder.java
@@ -88,7 +88,7 @@ public class DrugOrderBuilder {
         return this;
     }
 
-    public DrugOrderBuilder withDuration(double duration) {
+    public DrugOrderBuilder withDuration(int duration) {
         order.setDuration(duration);
         return this;
     }
